Automatically commit and push files once jobs are done
This change created a dependent slurm job to run after the first job
which created the day's date. The dependent job adds and commits the
data to a git repo for access by the RShiny app. This will put less
pressure on the biostat cluster filesystem than usng FTP within the
RShiny app.