Remove duplicate setting of today and data_repo from plots
This is already set in generate_plots which calls the plot scripts
using the source command. This means the objects are already in the
workspace. This change is made to remove an issue of some plot scripts
setting today incorrectly and others not setting it at all. This led
to plots using data from different days when run with the variable
production set to false.